Refs #124. Moved numbering testing to its own module
